requirements.txt
# Rail Madad Assistant — Python dependencies
# Pinned to minimum known-good versions

databricks-sdk>=0.28.0              # Databricks Workspace client — secrets, identity
databricks-sql-connector>=3.0.0     # SQL warehouse client — journey/ticket writes
databricks-vectorsearch>=0.40       # Mosaic AI Vector Search (RAG for chat)
fastapi>=0.110.0                    # Backend HTTP layer (Databricks App)
uvicorn[standard]>=0.27.0           # ASGI server for FastAPI
python-multipart>=0.0.9             # Multipart parser for /tickets/upload
pdfplumber>=0.11.0                  # PDF chunking for setup_vector_search.py
requests>=2.31.0                    # HTTP client for Sarvam / Tavily REST
streamlit>=1.30.0                   # Elderly-friendly grievance frontend (app.py)
